+++ core/initUI.lua
@@ -1,21 +1,22 @@
 local _, app = ...
 
 local COLORS = app.COLORS
+
 -- ========================
 -- Section: Merchant Frame build
 -- ========================
 local merchantFrameCost = app:frameBuilder({
     frameName = "CollectorHelper_Merchant",
     parent = MerchantFrame,
-    width = 280,
+    width = 330,
     height = MerchantFrame:GetHeight(),
     point = {
         pos = "TOPRIGHT",
-        x = 280,
+        x = 330,
         y = 0,
     },
     titleBuilder = {
-        text = app:textCFormat(COLORS.yellow, "Currency Needed to Collect this page"),
+        text = app:textCFormat(COLORS.yellow, "Currency Needed to Collect everything"),
         point = {
             pos = "TOP",
             x = 0,
@@ -24,24 +25,26 @@ local merchantFrameCost = app:frameBuilder({
     }
 })
 app.merchantFrameCost = merchantFrameCost
-app.marchantCost = app:fontBuilder({
+
+app.merchantCost = app:fontBuilder({
     parent = merchantFrameCost,
-    text = app:textCFormat(COLORS.green, "You have everything on this page"),
+    text = app:textCFormat(COLORS.green, "You have everything on this merchant"),
     point = {
         pos = "CENTER",
         x = 0,
         y = 80,
     }
 })
-app:buttonBuilder({
+
+local buyAllButton = app:buttonBuilder({
     buttonName = "Collector_MBuyButton",
     parent = merchantFrameCost,
     text = "Buy All Possible",
-    width = 120,
+    width = 100,
     height = 22,
     point = {
         pos = "BOTTOMLEFT",
-        x = 18,
+        x = 13,
         y = 8,
     },
     onClickScript = function(self, button)
@@ -55,15 +58,17 @@ app:buttonBuilder({
         end
     end
 })
-app:buttonBuilder({
+app.buyAllButton = buyAllButton
+
+local toggleShowOwned = app:buttonBuilder({
     buttonName = "Collector_MSHButton",
     parent = merchantFrameCost,
     text = "Show/hide owned",
-    width = 120,
+    width = 100,
     height = 22,
     point = {
-        pos = "BOTTOMRIGHT",
-        x = -18,
+        pos = "BOTTOM",
+        x = 0,
         y = 8,
     },
     onClickScript = function(self, button)
@@ -76,24 +81,58 @@ app:buttonBuilder({
         end
     end
 })
--- scrollable content cost
-app.mainScrollableContent = app:CreateScrollableContent(merchantFrameCost)
+app.toggleShowOwned = toggleShowOwned
 
--- ========================
--- Section: Merchant Source frame
--- ========================
-local merchantSourceItems = app:frameBuilder({
-    frameName = "CollectorHelper_MerchantSource",
+local ahListButton = app:buttonBuilder({
+    buttonName = "Collector_MAHListButton",
+    parent = merchantFrameCost,
+    text = "AH Track",
+    width = 100,
+    height = 22,
+    point = {
+        pos = "BOTTOMRIGHT",
+        x = -13,
+        y = 8,
+    },
+    onClickScript = function(self, button)
+        if button == "LeftButton" then
+            app.doAhTrack = true
+            app:updateShop()
+            app.ahFrame:Show()
+            app.ahScrollableContent.UpdateRows(app.ahItems)
+        end
+    end
+})
+app.ahListButton = ahListButton
+
+-- Scrollable content cost
+app.mainScrollableContent = app:CreateScrollableContent({
     parent = merchantFrameCost,
     width = 280,
-    height = MerchantFrame:GetHeight(),
+    height = 370,
     point = {
         pos = "TOPLEFT",
-        x = -280,
+        x = 8,
+        y = -37,
+    }
+})
+
+-- ========================
+-- Section: AH list
+-- ========================
+
+local ahFrame = app:frameBuilder({
+    frameName = "CollectorHelper_MerchantSource_AH",
+    parent = UIParent,
+    width = 330,
+    height = 400,
+    point = {
+        pos = "CENTER",
+        x = 0,
         y = 0,
     },
     titleBuilder = {
-        text = app:textCFormat(COLORS.yellow, "Possible Rewards for this Token"),
+        text = app:textCFormat(COLORS.yellow, "Collector Ah List"),
         point = {
             pos = "TOP",
             x = 0,
@@ -102,52 +141,83 @@ local merchantSourceItems = app:frameBuilder({
     }
 })
 
-app.merchantSourceItems = merchantSourceItems
-merchantSourceItems:Hide()
+-- Make the frame movable
+ahFrame:SetMovable(true)
+ahFrame:EnableMouse(true)
+ahFrame:RegisterForDrag("LeftButton")
+ahFrame:SetScript("OnDragStart", ahFrame.StartMoving)
 
-sourceDetails = app:fontBuilder({
-    parent = merchantSourceItems,
-    text = "",
+ahFrame:SetScript("OnDragStop", function(self)
+    self:StopMovingOrSizing()
+end)
+
+ahFrame:Hide()
+
+-- Scrollable content for AH list
+app.ahScrollableContent = app:CreateScrollableContent({
+    parent = ahFrame,
+    width = 280,
+    height = 320,
     point = {
         pos = "TOPLEFT",
-        x = 30,
-        y = -40,
+        x = 15,
+        y = -37,
     }
 })
 
+-- Close button for AH frame
 app:buttonBuilder({
-    buttonName = "Collector_MerchantSource_Close",
-    parent = merchantSourceItems,
+    buttonName = "Collector_MAHCloseButton",
+    parent = ahFrame,
     text = "Close",
-    width = 120,
+    width = 100,
     height = 22,
     point = {
-        pos = "BOTTOM",
-        x = 0,
+        pos = "BOTTOMRIGHT",
+        x = -13,
         y = 8,
     },
     onClickScript = function(self, button)
         if button == "LeftButton" then
-            merchantSourceItems:Hide()
-            merchantFrameCost:SetPoint("TOPRIGHT", 280, 0)
+            app.ahFrame:Hide()
         end
     end
 })
 
+-- Clear list button for AH frame
+app:buttonBuilder({
+    buttonName = "Collector_MAHClearButton",
+    parent = ahFrame,
+    text = "Clear List",
+    width = 100,
+    height = 22,
+    point = {
+        pos = "BOTTOMLEFT",
+        x = 13,
+        y = 8,
+    },
+    onClickScript = function(self, button)
+        if button == "LeftButton" then
+            app.ahItems = {}
+            app.ahScrollableContent.UpdateRows(app.ahItems)
+        end
+    end
+})
+app.ahFrame = ahFrame
 
 -- ========================
 -- Section: Other
 -- ========================
 app:buttonBuilder({
     buttonName = "Collector_MerchantToggle",
-    parent = MerchantFrame,
+    parent = _G["MerchantFrameTab1"],
     text = "Toggle CH",
-    width = 120,
-    height = 22,
+    width = 90,
+    height = 25,
     point = {
-        pos = "BOTTOM",
-        x = -100,
-        y = 5,
+        pos = "LEFT",
+        x = 3,
+        y = -28,
     },
     onClickScript = function(self, button)
         if button == "LeftButton" then