+++ core/itemOwnedCheck.lua
@@ -1,19 +1,19 @@
 local _, app = ...
 
-local _, playerclass = UnitClass("player")
---local token_Armor = TOKEN_ARMOR
-
+-- Check if the given item is owned by the player based on its type (toy, mount, pet, heirloom, etc.)
 function app:checkShopID(source)
-    -- 0 false, 1 true, 2 ignore
+    -- Return 0 for false, 1 for true, 2 for ignore
     if source == nil or source.itemId == nil then
         return 0
     end
-    -- CHECKS START
+
+    -- Check if the item is a toy and if the player already has it
     local isToy = app:isToyItem(source.itemId)
     if isToy then
         return PlayerHasToy(source.itemId) and 1 or 0
     end
 
+    -- Check if the item is a mount and if the player already knows it
     local isMount = app:IsMountItem(source.itemId)
     if isMount then
         local mountID = C_MountJournal.GetMountFromItem(source.itemId)
@@ -21,8 +21,9 @@ function app:checkShopID(source)
         return playerKnowsMount and 1 or 0
     end
 
-    local IsPet = app:IsPetItem(source.itemId)
-    if IsPet then
+    -- Check if the item is a pet and if the player already has it
+    local isPet = app:IsPetItem(source.itemId)
+    if isPet then
         local petName = C_PetJournal.GetPetInfoByItemID(source.itemId)
         if petName ~= nil then
             local _, petGUID = C_PetJournal.FindPetIDByName(petName)
@@ -30,6 +31,7 @@ function app:checkShopID(source)
         end
     end
 
+    -- Check if the item is part of a transmog set or an heirloom and if the player owns it
     local itemSetId = C_Item.GetItemLearnTransmogSet(source.itemId)
     local isItemHeirloom = C_Heirloom.IsItemHeirloom(source.itemId)
     if itemSetId ~= nil then
@@ -37,16 +39,13 @@ function app:checkShopID(source)
     elseif isItemHeirloom then
         return C_Heirloom.PlayerHasHeirloom(source.itemId) and 1 or 0
     else
+        -- Check if the item is part of the player's transmog collection
         local itemAppearanceID, _ = C_TransmogCollection.GetItemInfo(source.itemId)
         if itemAppearanceID == nil then
             local sourceID = app:GetSourceID(source.link)
             if sourceID ~= nil then
                 return C_TransmogCollection.PlayerHasTransmogItemModifiedAppearance(sourceID) and 1 or 0
             else
-                --local itemToken = token_Armor[source.itemId]
-                --if itemToken ~= nil then
-                --return app:tokenSourcesOwned(source.itemId, itemToken)
-                --end
                 return 2 -- ignore item
             end
         else
@@ -61,13 +60,12 @@ function app:checkShopID(source)
     end
 end
 
--- Item is BoE check if already purchased and is in bags
+-- Check if the item is in the player's bags
 function app:playerHasItemInBag(itemCheckId)
-    -- check if item is in bag
     for bag = 0, NUM_BAG_SLOTS do
         for slot = 1, C_Container.GetContainerNumSlots(bag) do
             local item = C_Container.GetContainerItemInfo(bag, slot)
-            if (item ~= nil and item.itemID == itemCheckId) then
+            if item ~= nil and item.itemID == itemCheckId then
                 return true
             end
         end
@@ -75,31 +73,22 @@ function app:playerHasItemInBag(itemCheckId)
     return false
 end
 
--- TOY CHECK
+-- Check if the item is a toy
 function app:isToyItem(itemId)
-    if C_ToyBox.GetToyInfo(itemId) then
-        return true
-    end
-    return false
+    return C_ToyBox.GetToyInfo(itemId) ~= nil
 end
 
--- MOUNT CHECK
+-- Check if the item is a mount
 function app:IsMountItem(itemId)
-    if C_MountJournal.GetMountFromItem(itemId) then
-        return true
-    end
-    return false
+    return C_MountJournal.GetMountFromItem(itemId) ~= nil
 end
 
--- PET CHECK
+-- Check if the item is a pet
 function app:IsPetItem(itemId)
-    if C_PetJournal.GetPetInfoByItemID(itemId) then
-        return true
-    end
-    return false
+    return C_PetJournal.GetPetInfoByItemID(itemId) ~= nil
 end
 
--- SET CHECK
+-- Check if all items in the transmog set are owned by the player
 function app:setOwned(itemSet)
     local ids = C_TransmogSets.GetAllSourceIDs(itemSet)
     for _, sourceID in ipairs(ids) do
@@ -107,36 +96,4 @@ function app:setOwned(itemSet)
         if not info.isCollected then return false end
     end
     return true
-end
-
--- In work for next update
--- Token item that provides specific items class based
-function app:tokenSourcesOwned(tokenId, itemToken)
-    local function isItemValidForClass(classes)
-        for _, class in pairs(classes) do
-            if class == app.CLS.ANY then
-                -- can be used by any class
-                return true
-            end
-            if class == playerclass then
-                return true
-            end
-        end
-        return false
-    end
-
-    local count = 0
-    local owned = 0
-    app.merchantTokenMap[tokenId] = {}
-    for itemId, classes in pairs(itemToken) do
-        if (isItemValidForClass(classes)) then
-            count = count + 1
-            local isOwned = C_TransmogCollection.PlayerHasTransmog(itemId)
-            if isOwned then
-                owned = owned + 1
-            end
-            table.insert(app.merchantTokenMap[tokenId], { owned = isOwned, itemId = itemId })
-        end
-    end
-    return (owned == count) and 1 or 15
 end